refactor: separate client from spec for secrets pkg
